Removed AsyncStatus enum & moved LocalStore enum
