+++ src/OfflinePin.tsx
@@ -2,10 +2,9 @@ import { Button, Input, Label } from '@equinor/eds-core-react';
 import { Navbar, ProcosysButton } from '@equinor/procosys-webapp-components';
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import { LocalStorage } from './contexts/McAppContext';
 import { db } from './offline/db';
 import { updateOfflineStatus } from './offline/OfflineStatus';
-import { OfflineStatus } from './typings/enums';
+import { LocalStorage, OfflineStatus } from './typings/enums';
 
 const ContentWrapper = styled.main`
     margin: 16px;
