+++ src/components/navigation/SideMenu.tsx
@@ -6,9 +6,8 @@ import PlantContext from '../../contexts/PlantContext';
 import useCommonHooks from '../../utils/useCommonHooks';
 import { COLORS } from '../../style/GlobalStyles';
 import { isOfType, StorageKey } from '@equinor/procosys-webapp-components';
-import { OfflineStatus } from '../../typings/enums';
+import { LocalStorage, OfflineStatus } from '../../typings/enums';
 import packageJson from '../../../package.json';
-import { LocalStorage } from '../../contexts/McAppContext';
 
 const SideMenuWrapper = styled.aside<{ isActive: boolean }>`
     width: 297px;
