+++ src/offline/OfflineStatus.ts
@@ -1,6 +1,5 @@
 import { isOfType } from '@equinor/procosys-webapp-components';
-import { LocalStorage } from '../contexts/McAppContext';
-import { OfflineStatus } from '../typings/enums';
+import { LocalStorage, OfflineStatus } from '../typings/enums';
 
 export const getOfflineStatusfromLocalStorage = (): OfflineStatus => {
     const offlineStatus = localStorage.getItem(LocalStorage.OFFLINE_STATUS);
