+++ src/pages/Entity/WorkOrderInfo.tsx
@@ -1,6 +1,7 @@
 import React, { useContext } from 'react';
 import { Button } from '@equinor/eds-core-react';
 import {
+    AsyncStatus,
     Attachments,
     CollapsibleCard,
     ErrorPage,
@@ -9,7 +10,6 @@ import {
 } from '@equinor/procosys-webapp-components';
 import styled from 'styled-components';
 import AsyncPage from '../../components/AsyncPage';
-import { AsyncStatus } from '../../contexts/McAppContext';
 import { isOfType } from '../../services/apiTypeGuards';
 import {
     Attachment,
