+++ src/pages/Search/SavedSearches/SavedSearchResult.tsx
@@ -2,10 +2,10 @@ import { Button } from '@equinor/eds-core-react';
 import React from 'react';
 import styled from 'styled-components';
 import EdsIcon from '../../../components/icons/EdsIcon';
-import { AsyncStatus } from '../../../contexts/McAppContext';
 import { SavedSearch, ApiSavedSearchType } from '../../../services/apiTypes';
 import { Caption, COLORS } from '../../../style/GlobalStyles';
 import useCommonHooks from '../../../utils/useCommonHooks';
+import { AsyncStatus } from '@equinor/procosys-webapp-components';
 
 const SavedSearchWrapper = styled.article`
     cursor: pointer;
