+++ src/pages/SelectPlant/SelectPlant.test.tsx
@@ -1,10 +1,10 @@
 import { render } from '@testing-library/react';
 import React from 'react';
-import { AsyncStatus } from '../../contexts/McAppContext';
 import { withMcAppContext } from '../../test/contexts';
 import { testPlants } from '../../test/dummyData';
 import { OfflineStatus } from '../../typings/enums';
 import SelectPlant from './SelectPlant';
+import { AsyncStatus } from '@equinor/procosys-webapp-components';
 
 describe('<SelectPlant />', () => {
     it('Renders plant buttons successfully upon loading', () => {
