+++ src/pages/SelectProject/SelectProject.test.tsx
@@ -1,10 +1,10 @@
 import { render } from '@testing-library/react';
 import React from 'react';
-import { AsyncStatus } from '../../contexts/McAppContext';
 import { withPlantContext } from '../../test/contexts';
 import { testProjects } from '../../test/dummyData';
 import { OfflineStatus } from '../../typings/enums';
 import SelectProject from '../SelectProject/SelectProject';
+import { AsyncStatus } from '@equinor/procosys-webapp-components';
 
 describe('<SelectProject />', () => {
     it('Renders project buttons successfully upon loading', () => {
